# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Montor-ing API",
        "description": "Dokumentasi API untuk aplikasi Montor-ing (Inventory & Sales Monitoring).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Server API Utama"
        }
    ],
    "paths": {
...
        "/master/divisions": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar divisi",
                "description": "Mengambil data devisi",
                "operationId": "748e77e2dd48ac2957da97697e9ff4ca",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/master/departments": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar departemen",
                "description": "Mengambil data departemen",
                "operationId": "b66eca7505bf1addc18450d0da1e6600",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    },
                    {
                        "name": "division_id",
                        "in": "query",
                        "description": "ID divisi",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/master/suppliers": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar supplier",
                "description": "Mengambil data supplier",
                "operationId": "81d72ce5d556a98593d1b8bacc768c0f",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/master/product-groups": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar group produk",
                "description": "Mengambil data group produk",
                "operationId": "bf53f8d66a73181b6f09a099dd8a9a57",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/master/product-brands": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar brand produk",
                "description": "Mengambil data brand produk",
                "operationId": "859da4d9f0a7d1ed571214c936937c14",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/master/accounts": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar akun",
                "description": "Mengambil data akun",
                "operationId": "54a43c885142576d16c1b14ba52c9dee",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/master/user-config-rules": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar user config rules",
                "description": "Mengambil data user config rules",
                "operationId": "3ed44cb701f218cecf92d0950b824994",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/master/customer-groups": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar group pelanggan",
                "description": "Mengambil data group pelanggan",
                "operationId": "d2856e558b6d5ed0ad0b5d61ab574e92",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Pencarian Nama Group Pelanggan",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/master/customers": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar pelanggan",
                "description": "Mengambil data pelanggan",
                "operationId": "153908b2076b53dbf9d2ad7048bd567a",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    },
                    {
                        "name": "customergroup_id",
                        "in": "query",
                        "description": "ID Group Pelanggan",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Pencarian Nama Pelanggan",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/master/company": {
            "get": {
                "tags": [
                    "Master"
                ],
                "summary": "Daftar perusahaan",
                "description": "Mengambil data perusahaan",
                "operationId": "194cdc1d3b1d6949dc3f62b197badb60",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/X-Database-Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sukses",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "parameters": {
            "X-Database-Name": {
                "name": "X-Database-Name",
                "in": "header",
                "description": "Nama database yang digunakan",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token dari endpoint /login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}